pyproject.toml
[tool.poetry]
name = "ramanchada-api"
version = "0.1.0"
description = "A web API for the RamanChada 2 Raman spectroscopy harmonisation library, incorporating the AMBIT/eNanoMapper data model"
authors = [
  "Nina Jeliazkova <jeliazkova.nina@gmail.com>",
]
maintainers = [
  "Luchesar Iliev <luchesar.iliev@gmail.com>",
  "Vedrin Jeliazkov <vedrin.jeliazkov@gmail.com>",
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/h2020charisma/ramanchada-api"
packages = [
  { include = "rcapi", from = "src" },
]
include = [
  "src/rcapi/config/config.yaml",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apscheduler = "^3.10.1"
fastapi = "^0.112.1"
h5grove = "^2.0.0"
h5pyd = { git = "https://github.com/HDFGroup/h5pyd.git" }
numcompress = "^0.1.2"
openpyxl = "^3.1.2"
pydantic-settings = "^2.4.0"
pynanomapper = { path = "extern/pynanomapper", develop = true }
python-keycloak = "^4.3.0"
python-multipart = "^0.0.6"
ramanchada2 = { path = "extern/ramanchada2", develop = true }
uvicorn = "^0.27.1"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/h2020charisma/ramanchada-api/issues"

[tool.poetry.scripts]
dev = "scripts.dev_server:main"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
pytest = "^8.0.1"

[tool.poetry.group.jupyter.dependencies]
jupyter = "^1.0.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
